Put Neotree on the right side and enable toggling
Neotree being on the right side is just a personal preference.

The toggle option is quite nice because often I switch the focus
manually back to the editor and then, using the keymap for Neotree with
the intention to close it, it actually puts the focus to Neotree, where
I need to use the keymap *again* to actually close it. The toggle option
allows to close Neotree from anymore if already opened.